requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.1
Automat==20.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
blessed==1.17.12
boto3==1.16.26
botocore==1.19.26
cached-property==1.5.2
cement==2.8.2
certifi==2020.11.8
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
coolsms-python-sdk==2.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
daphne==3.0.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.5.0
django-crontab==0.7.1
django-filter==2.4.0
django-hosts==4.0
django-rest-auth==0.9.5
django-restframework==0.0.1
django-ses==1.0.3
django-storages==1.10.1
djangorestframework==3.11.0
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.6.0
docker==4.4.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
future==0.16.0
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==3.1.1
incremental==17.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==1.1.1
msgpack==1.0.1
oauthlib==3.1.0
packaging==20.7
paramiko==2.7.2
pathspec==0.5.9
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==20.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
semantic-version==2.5.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.57.0
zipp==3.4.0
zope.interface==5.2.0
